requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
pydantic
python-multipart